Extract common function
